requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
distlib==0.3.9
dj-database-url==3.0.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
idna==3.10
platformdirs==4.3.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pygame==2.6.1
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
